新增遊戲起始畫面、結算畫面(美術待補) 1. 修改部分函式名稱。 2. image2.png->image.png。 3. 遊戲啟動時在選單畫面、遊戲結束跳轉到結算畫面。
