+++ G9_10/src/G9_10.java
@@ -3,19 +3,19 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/javafx/FXMain.java to edit this template
  */
 
- import javafx.application.Application;
- import javafx.fxml.FXMLLoader;
- import javafx.scene.Parent;
- import javafx.scene.Scene;
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
  
  public class G9_10 extends Application {
      @Override
      public void start(Stage stage) throws Exception {
-         Parent root = FXMLLoader.load(getClass().getResource("FXML.fxml"));
+         Parent root = FXMLLoader.load(getClass().getResource("MenuController.fxml"));
          Scene scene = new Scene(root);
-         Image img =  new Image("icon.jpg"); //縮圖及背景圖片
+         Image img = new Image(getClass().getResource("icon.jpg").toExternalForm());//縮圖及背景圖片
          stage.getIcons().add(img);
          stage.setTitle("圓射!啟動!!");
          stage.setScene(scene);
@@ -25,4 +25,4 @@ import javafx.stage.Stage;
      public static void main(String[] args) {
          launch(args);
      }
- }
+ }
