+++ G9_10/src/GameOverController.fxml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Font?>
+<?import javafx.scene.text.Text?>
+
+
+<AnchorPane maxHeight="466.0" maxWidth="364.0" minHeight="466.0" minWidth="364.0" prefHeight="466.0" prefWidth="364.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="GameOverController">
+   <children>
+      <Label fx:id="label_final_score" layoutX="92.0" layoutY="84.0" prefHeight="56.0" prefWidth="272.0">
+         <font>
+            <Font size="35.0" />
+         </font>
+      </Label>
+      <Text layoutX="92.0" layoutY="69.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Final Score">
+         <font>
+            <Font size="35.0" />
+         </font>
+      </Text>
+      <Button fx:id="button_play_again" ellipsisString="" layoutX="92.0" layoutY="259.0" mnemonicParsing="false" onAction="#Click_Play_Again" prefHeight="80.0" prefWidth="181.0" text="Play Again">
+         <font>
+            <Font size="28.0" />
+         </font>
+      </Button>
+      <Button fx:id="button_title" ellipsisString="" layoutX="92.0" layoutY="356.0" mnemonicParsing="false" onAction="#Click_title" prefHeight="80.0" prefWidth="181.0" text="Title">
+         <font>
+            <Font size="28.0" />
+         </font>
+      </Button>
+      <Text layoutX="92.0" layoutY="181.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Combo">
+         <font>
+            <Font size="35.0" />
+         </font>
+      </Text>
+      <Label fx:id="label_combo" layoutX="92.0" layoutY="196.0" prefHeight="56.0" prefWidth="272.0">
+         <font>
+            <Font size="35.0" />
+         </font>
+      </Label>
+   </children>
+</AnchorPane>
